@@ -6,42 +6,37 @@ git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 source "https://rubygems.org"
 git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-ruby "3.1.2"
+ruby "3.2.2"
 
 # Bundle edge Rails instead: gem 'rails', [https://github.com/rails/rails]
-gem "rails", "7.0.4"
+gem "rails", "7.0.7.2"
 
 # The original asset pipeline for Rails [https://github.com/rails/sprockets-rails]
 gem "sprockets-rails", "3.4.2"
 
 # Use postgresql as the database for Active Record
-gem "pg", "1.2.3"
+gem "pg"
 
 # Use Puma as the app server [https://github.com/puma/puma]
-gem "puma", "6.0.0"
+gem "puma", "6.3.1"
 
 # Bundle and transpile JavaScript [https://github.com/rails/jsbundling-rails]
-gem "jsbundling-rails", "1.0.3"
+gem "jsbundling-rails", "1.1.2"
 
 # Bundle and process CSS [https://github.com/rails/cssbundling-rails]
-gem "cssbundling-rails", "1.1.1"
+gem "cssbundling-rails", "1.2.0"
 
 # Hotwire's SPA-like page accelerator [https://github.com/hotwired/turbo-rails]
-gem "turbo-rails", "1.3.2"
+gem "turbo-rails", "1.4.0"
 
 # Hotwire's modest JavaScript framework [https://github.com/hotwired/stimulus-rails]
-gem "stimulus-rails", "1.1.1"
+gem "stimulus-rails", "1.2.2"
 
-# Build reactive applications [https://github.com/stimulusreflex/stimulus_reflex]
-gem "stimulus_reflex", "= 3.5.0.pre8"
-
-
+gem "devise"
 gem "devise_invitable", "~> 2.0.0"
-# Use postgresql as the database for Active Record
+gem "devise-pwned_password"
 
 gem "draper"
-gem "devise", github: "heartcombo/devise", branch: "main"
-gem "devise-pwned_password"
 gem "letter_opener_web", group: :development
 gem "pundit"
 # Use Active Storage variant
@@ -49,18 +44,19 @@ gem "pundit"
 gem "wicked_pdf", github: "mileszs/wicked_pdf", branch: "master"
 gem "wkhtmltopdf-binary"
 gem "hiredis"
-gem "redis", "4.8.0"
+gem "redis", "5.0.7"
 gem "json"
 gem "pagy"
 
 # sidekiq gems, sinatra is used to build UI for /sidekiq
-gem "sidekiq", "6.5.8"
-gem "sidekiq-scheduler", "4.0.2"
+gem "sidekiq", "7.1.2"
+gem "sidekiq-scheduler", "5.0.3"
 gem "sinatra", ">= 1.3.0", require: nil
 
 # Reduces boot times through caching; required in config/boot.rb
 gem "acts_as_tenant"
-gem "bootsnap", "1.13.0", require: false
+gem "bootsnap", "1.16.0", require: false
+gem "rails-patterns"
 gem "rexml"
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
@@ -72,10 +68,10 @@ group :development do
   gem "web-console", ">= 4.1.0"
   # Display performance information such as SQL time and flame graphs for each request in your browser.
   # Can be configured to work on production as well see: https://github.com/MiniProfiler/rack-mini-profiler/blob/master/README.md
-  gem "rack-mini-profiler", "~> 2.0"
+  gem "rack-mini-profiler", "~> 3.1"
   gem "listen", "~> 3.3"
   # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
-  gem "spring"
+  #   gem "spring"
 end
 
 group :test do
