@@ -5,13 +5,13 @@ class ApplicationController < ActionController::Base
   before_action :configure_permitted_parameters, if: :devise_controller?
 
   include Pagy::Backend
-  include Pundit
+  include Pundit::Authorization
 
   rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
-  rescue_from ActiveRecord::RecordNotFound, with: :record_not_found
-  rescue_from ActionController::InvalidAuthenticityToken, with: :invalid_token
   rescue_from Pundit::NotDefinedError, with: :record_not_found
+  rescue_from ActiveRecord::RecordNotFound, with: :record_not_found
   rescue_from ActiveRecord::InvalidForeignKey, with: :show_referenced_alert
+  rescue_from ActionController::InvalidAuthenticityToken, with: :invalid_token
 
   LIMIT = 30
 
@@ -22,29 +22,45 @@ class ApplicationController < ActionController::Base
     Current.account = current_user.account if current_user
   end
 
-  etag {
+  etag do
     if Rails.env == "production" or Rails.env == "staging"
       heroku_version
     else
-      "screener"
+      'screener'
     end
-  }
+  end
 
   fragment_cache_key do
-    "screener"
+    'screener'
   end
 
   def heroku_version
-    ENV["HEROKU_RELEASE_VERSION"] if Rails.env == "production" or Rails.env == "staging"
+    ENV['HEROKU_RELEASE_VERSION'] if Rails.env == 'production' or Rails.env == 'staging'
   end
 
   def render_partial(partial, collection:, cached: true)
     respond_to do |format|
       format.html
-      format.json {
-        render json: { entries: render_to_string(partial: partial, formats: [:html], collection: collection, cached: cached),
-                       pagination: render_to_string(partial: "shared/paginator", formats: [:html], locals: { pagy: @pagy }) }
-      }
+      format.json do
+        render json: { entries: render_to_string(partial:, formats: [:html], collection:, cached:),
+                       pagination: render_to_string(partial: 'shared/paginator', formats: [:html],
+                                                    locals: { pagy: @pagy }) }
+      end
+    end
+  end
+
+  def after_invite_path_for(_resource)
+    users_path
+  end
+
+  def render_partial_as(partial, collection:, as:, cached: true)
+    respond_to do |format|
+      format.html
+      format.json do
+        render json: { entries: render_to_string(partial:, formats: [:html], collection:, as:, cached:),
+                       pagination: render_to_string(partial: 'shared/paginator', formats: [:html],
+                                                    locals: { pagy: @pagy }) }
+      end
     end
   end
 
@@ -52,23 +68,25 @@ class ApplicationController < ActionController::Base
     @redirect_path = request.path
   end
 
-  def show_referenced_alert(exception)
+  def show_referenced_alert(_exception)
     respond_to do |format|
-      format.turbo_stream {
-        render turbo_stream: turbo_stream.replace("modal", partial: "shared/modal", locals: { title: "Unable to Delete Record", message: "This record has been associated with other records in system therefore deleting this might result in unexpected behavior. If you want to delete this please make sure all assosications have been removed first.", main_button_visible: false })
-      }
+      format.turbo_stream do
+        render turbo_stream: turbo_stream.replace('modal', partial: 'shared/modal',
+                                                           locals: { title: 'Unable to Delete Record', message: 'This record has been associated with other records in system therefore deleting this might result in unexpected behavior. If you want to delete this please make sure all assosications have been removed first.', main_button_visible: false })
+      end
     end
   end
 
   def show_delete_confirmation_alert
-    show_confirmation_alert("Delete Record", "Are you sure you want to delete this record?")
+    show_confirmation_alert('Delete Record', 'Are you sure you want to delete this record?')
   end
 
   def show_confirmation_alert(title, message)
     respond_to do |format|
-      format.turbo_stream {
-        render turbo_stream: turbo_stream.replace("modal", partial: "shared/modal", locals: { title: title, message: message, main_button_visible: true })
-      }
+      format.turbo_stream do
+        render turbo_stream: turbo_stream.replace('modal', partial: 'shared/modal',
+                                                           locals: { title:, message:, main_button_visible: true })
+      end
     end
   end
 
@@ -82,7 +100,7 @@ class ApplicationController < ActionController::Base
     end
   end
 
-  def after_sign_out_path_for(resource)
+  def after_sign_out_path_for(_resource)
     root_path
   end
 
@@ -90,7 +108,7 @@ class ApplicationController < ActionController::Base
     redirect_to(root_path)
   end
 
-  def signed_in_root_path(resource)
+  def signed_in_root_path(_resource)
     root_path
   end
 
@@ -104,28 +122,30 @@ class ApplicationController < ActionController::Base
 
   def invalid_token
     sign_out(current_user) if current_user
-    redirect_to new_user_session_path, alert: "Your session has expired. Please login again."
+    redirect_to new_user_session_path, alert: 'Your session has expired. Please login again.'
   end
 
   def render_timeline(partial, collection:, cached: true)
     respond_to do |format|
       format.html
-      format.json {
-        render json: { entries: render_to_string(partial: partial, formats: [:html], collection: collection, as: :event, cached: cached),
-                       pagination: render_to_string(partial: "shared/paginator", formats: [:html], locals: { pagy: @pagy }) }
-      }
+      format.json do
+        render json: { entries: render_to_string(partial:, formats: [:html], collection:, as: :event, cached:),
+                       pagination: render_to_string(partial: 'shared/paginator', formats: [:html],
+                                                    locals: { pagy: @pagy }) }
+      end
     end
   end
 
   def pagy_nil_safe(params, collection, vars = {})
     pagy = Pagy.new(count: collection.count(:all), page: params[:page], **vars)
     return pagy, collection.offset(pagy.offset).limit(pagy.items) if collection.respond_to?(:offset)
-    return pagy, collection
+
+    [pagy, collection]
   end
 
   protected
 
   def configure_permitted_parameters
-    devise_parameter_sanitizer.permit(:accept_invitation, keys: [:first_name, :last_name])
+    devise_parameter_sanitizer.permit(:accept_invitation, keys: %i[first_name last_name])
   end
 end
