@@ -1,10 +1,10 @@
-import { Application } from "@hotwired/stimulus"
+import { Application } from "@hotwired/stimulus";
 
-const application = Application.start()
+const application = Application.start();
 
 // Configure Stimulus development experience
-application.warnings = true
-application.debug    = false
-window.Stimulus      = application
+application.warnings = true;
+application.debug = false;
+window.Stimulus = application;
 
-export { application }
+export { application };
