@@ -1,5 +1,4 @@
-import { Controller } from 'stimulus'
-import StimulusReflex from 'stimulus_reflex'
+import { Controller } from "@hotwired/stimulus";
 
 /* This is your ApplicationController.
  * All StimulusReflex controllers should inherit from this class.
@@ -13,10 +12,6 @@ import StimulusReflex from 'stimulus_reflex'
  * Learn more at: https://docs.stimulusreflex.com
  */
 export default class extends Controller {
-  connect () {
-    StimulusReflex.register(this)
-  }
-
   /* Application-wide lifecycle methods
    *
    * Use these methods to handle lifecycle concerns for the entire application.
@@ -34,27 +29,27 @@ export default class extends Controller {
    *   reflexId - a UUID4 or developer-provided unique identifier for each Reflex
    */
 
-  beforeReflex (element, reflex, noop, reflexId) {
+  beforeReflex(element, reflex, noop, reflexId) {
     // document.body.classList.add('wait')
   }
 
-  reflexSuccess (element, reflex, noop, reflexId) {
+  reflexSuccess(element, reflex, noop, reflexId) {
     // show success message
   }
 
-  reflexError (element, reflex, error, reflexId) {
+  reflexError(element, reflex, error, reflexId) {
     // show error message
   }
 
-  reflexHalted (element, reflex, error, reflexId) {
+  reflexHalted(element, reflex, error, reflexId) {
     // handle aborted Reflex action
   }
 
-  afterReflex (element, reflex, noop, reflexId) {
+  afterReflex(element, reflex, noop, reflexId) {
     // document.body.classList.remove('wait')
   }
 
-  finalizeReflex (element, reflex, noop, reflexId) {
+  finalizeReflex(element, reflex, noop, reflexId) {
     // all operations have completed, animation etc is now safe
   }
 }
