@@ -1,12 +1,12 @@
-import { Controller } from "@hotwired/stimulus"
+import { Controller } from "@hotwired/stimulus";
 
 export default class extends Controller {
- static values = { message: String }
+  static values = { message: String };
 
   confirm(event) {
-    if (!(window.confirm(this.messageValue))) {
-      event.preventDefault()
-      event.stopImmediatePropagation()
+    if (!window.confirm(this.messageValue)) {
+      event.preventDefault();
+      event.stopImmediatePropagation();
     }
   }
 }
