@@ -1,32 +1,19 @@
-// This file is auto-generated by ./bin/rails stimulus:manifest:update
-// Run that command whenever you add a new controller
+import { application } from "./application";
 
-import { application } from "./application"
+import controllers from "./**/*_controller.js";
 
-import ApplicationController from "./application_controller"
-application.register("application", ApplicationController)
+controllers.forEach((controller) => {
+  application.register(controller.name, controller.module.default);
+});
 
-import DropdownController from "./dropdown_controller"
-application.register("dropdown", DropdownController)
+import StimulusSlimSelect from "./slim_select_controller";
+application.register("slimselect", StimulusSlimSelect);
 
-import InfiniteScrollController from "./infinite_scroll_controller"
-application.register("infinite-scroll", InfiniteScrollController)
+import Slideover from "./slideover_controller";
+application.register("slideover", Slideover);
 
-import ModalController from "./modal_controller"
-application.register("modal", ModalController)
+import Attempt from "./attempt_controller";
+application.register("attempt", Attempt);
 
-import NavSearchController from "./nav_search_controller"
-application.register("nav-search", NavSearchController)
-
-import ConfirmationController from "./confirmation_controller"
-import StimulusReflex from 'stimulus_reflex'
-import consumer from '../channels/consumer'
-import controller from '../controllers/application_controller'
-import CableReady from 'cable_ready'
-application.register("confirmation", ConfirmationController)
-application.consumer = consumer
-StimulusReflex.initialize(application, { controller, isolate: true })
-CableReady.initialize({ consumer })
- application.consumer = consumer
-application.consumer = consumer
-application.consumer = consumer
+import DragItem from "./drag_item_controller";
+application.register("drag-item", DragItem);
