@@ -1,13 +1,11 @@
-import { Controller } from "@hotwired/stimulus"
+import { Controller } from "@hotwired/stimulus";
 
 export default class extends Controller {
+  close() {
+    this.element.remove();
+  }
 
-    close() {
-        this.element.remove()
-    }
-
-    escClose(event) {
-        if (event.key === 'Escape') this.close()
-    }
-
+  escClose(event) {
+    if (event.key === "Escape") this.close();
+  }
 }
