@@ -1,5 +1,5 @@
 class Screening::HrPolicy < Screening::BaseScreeningPolicy
   def index?
-    user.admin? || user.hr? || user.team_lead?
+    true
   end
 end
