@@ -1,5 +1,5 @@
 class Screening::InterviewPolicy < Screening::BaseScreeningPolicy
   def index?
-    user.admin? || user.interviewer? || user.team_lead?
+    true
   end
 end
