@@ -1,5 +1,5 @@
 class Screening::ResumePolicy < Screening::BaseScreeningPolicy
   def index?
-    user.admin? || user.resume_screener? || user.team_lead?
+    true
   end
 end
