@@ -1,5 +1,5 @@
 class Screening::TelephonicPolicy < Screening::BaseScreeningPolicy
   def index?
-    user.admin? || user.telephonic_screener? || user.team_lead?
+    true
   end
 end
