@@ -10,7 +10,11 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
+<<<<<<< HEAD
 ActiveRecord::Schema[7.0].define(version: 2023_01_13_104231) do
+=======
+ActiveRecord::Schema[7.0].define(version: 2023_08_29_055601) do
+>>>>>>> dev
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
@@ -25,8 +29,14 @@ ActiveRecord::Schema[7.0].define(version: 2023_01_13_104231) do
     t.text "body"
     t.string "record_type", null: false
     t.bigint "record_id", null: false
+<<<<<<< HEAD
     t.datetime "created_at", precision: nil, null: false
     t.datetime "updated_at", precision: nil, null: false
+=======
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["record_type", "record_id", "name"], name: "index_action_text_rich_texts_uniqueness", unique: true
+>>>>>>> dev
   end
 
   create_table "active_storage_attachments", force: :cascade do |t|
@@ -34,7 +44,13 @@ ActiveRecord::Schema[7.0].define(version: 2023_01_13_104231) do
     t.string "record_type", null: false
     t.bigint "record_id", null: false
     t.bigint "blob_id", null: false
+<<<<<<< HEAD
     t.datetime "created_at", precision: nil, null: false
+=======
+    t.datetime "created_at", null: false
+    t.index ["blob_id"], name: "index_active_storage_attachments_on_blob_id"
+    t.index ["record_type", "record_id", "name", "blob_id"], name: "index_active_storage_attachments_uniqueness", unique: true
+>>>>>>> dev
   end
 
   create_table "active_storage_blobs", force: :cascade do |t|
@@ -45,7 +61,12 @@ ActiveRecord::Schema[7.0].define(version: 2023_01_13_104231) do
     t.string "service_name", null: false
     t.bigint "byte_size", null: false
     t.string "checksum"
+<<<<<<< HEAD
     t.datetime "created_at", precision: nil, null: false
+=======
+    t.datetime "created_at", null: false
+    t.index ["key"], name: "index_active_storage_blobs_on_key", unique: true
+>>>>>>> dev
   end
 
   create_table "active_storage_variant_records", force: :cascade do |t|
@@ -92,13 +113,61 @@ ActiveRecord::Schema[7.0].define(version: 2023_01_13_104231) do
     t.datetime "updated_at", null: false
   end
 
+  create_table "folders", force: :cascade do |t|
+    t.string "title"
+    t.bigint "user_id", null: false
+    t.bigint "space_id", null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["space_id"], name: "index_folders_on_space_id"
+    t.index ["user_id"], name: "index_folders_on_user_id"
+  end
+
+  create_table "pinned_spaces", force: :cascade do |t|
+    t.bigint "user_id", null: false
+    t.bigint "space_id", null: false
+    t.index ["space_id"], name: "index_pinned_spaces_on_space_id"
+    t.index ["user_id"], name: "index_pinned_spaces_on_user_id"
+  end
+
+  create_table "recent_surveys", force: :cascade do |t|
+    t.bigint "user_id", null: false
+    t.bigint "survey_surveys_id", null: false
+    t.integer "count", default: 0
+    t.index ["survey_surveys_id"], name: "index_recent_surveys_on_survey_surveys_id"
+    t.index ["user_id"], name: "index_recent_surveys_on_user_id"
+  end
+
+  create_table "spaces", force: :cascade do |t|
+    t.string "title"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.bigint "user_id"
+    t.string "description"
+    t.boolean "archive", default: false
+    t.datetime "archive_at"
+    t.index ["user_id"], name: "index_spaces_on_user_id"
+  end
+
+  create_table "spaces_users", id: false, force: :cascade do |t|
+    t.bigint "user_id", null: false
+    t.bigint "space_id", null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+  end
+
   create_table "survey_answers", force: :cascade do |t|
     t.integer "attempt_id"
     t.integer "question_id"
     t.integer "option_id"
     t.boolean "correct"
+<<<<<<< HEAD
     t.datetime "created_at", precision: nil, null: false
     t.datetime "updated_at", precision: nil, null: false
+=======
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+>>>>>>> dev
     t.integer "score", default: 0, null: false
   end
 
@@ -108,8 +177,13 @@ ActiveRecord::Schema[7.0].define(version: 2023_01_13_104231) do
     t.boolean "winner"
     t.integer "score"
     t.string "comment"
+<<<<<<< HEAD
     t.datetime "created_at", precision: nil, default: "2021-12-24 12:55:04", null: false
     t.datetime "updated_at", precision: nil, default: "2021-12-24 12:55:04", null: false
+=======
+    t.datetime "created_at", default: "2021-12-24 12:01:17", null: false
+    t.datetime "updated_at", default: "2021-12-24 12:01:17", null: false
+>>>>>>> dev
     t.integer "actor_id", null: false
   end
 
@@ -118,23 +192,39 @@ ActiveRecord::Schema[7.0].define(version: 2023_01_13_104231) do
     t.integer "weight", default: 0
     t.string "text"
     t.boolean "correct"
+<<<<<<< HEAD
     t.datetime "created_at", precision: nil, null: false
     t.datetime "updated_at", precision: nil, null: false
+=======
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+>>>>>>> dev
   end
 
   create_table "survey_participant", force: :cascade do |t|
     t.string "name"
     t.string "email"
+<<<<<<< HEAD
     t.datetime "created_at", precision: nil, null: false
     t.datetime "updated_at", precision: nil, null: false
+=======
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+>>>>>>> dev
   end
 
   create_table "survey_questions", force: :cascade do |t|
     t.integer "survey_id"
     t.string "text"
     t.string "description"
+<<<<<<< HEAD
     t.datetime "created_at", precision: nil, null: false
     t.datetime "updated_at", precision: nil, null: false
+=======
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.integer "order", default: 0, null: false
+>>>>>>> dev
   end
 
   create_table "survey_surveys", force: :cascade do |t|
@@ -142,11 +232,22 @@ ActiveRecord::Schema[7.0].define(version: 2023_01_13_104231) do
     t.text "description"
     t.integer "attempts_number", default: 0
     t.boolean "finished", default: false
-    t.boolean "active", default: false
+    t.boolean "active", default: true
     t.integer "winning_score", default: 0
     t.integer "survey_type", default: 0
+<<<<<<< HEAD
     t.datetime "created_at", precision: nil, null: false
     t.datetime "updated_at", precision: nil, null: false
+=======
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.integer "survey_stage"
+    t.bigint "folder_id"
+    t.date "archived_on"
+    t.integer "user_id"
+    t.boolean "pin", default: false
+    t.index ["folder_id"], name: "index_survey_surveys_on_folder_id"
+>>>>>>> dev
   end
 
   create_table "users", force: :cascade do |t|
@@ -155,6 +256,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_01_13_104231) do
     t.string "email", default: "", null: false
     t.string "encrypted_password", default: "", null: false
     t.string "reset_password_token"
+<<<<<<< HEAD
     t.datetime "reset_password_sent_at", precision: nil
     t.datetime "remember_created_at", precision: nil
     t.datetime "created_at", precision: nil, null: false
@@ -167,10 +269,25 @@ ActiveRecord::Schema[7.0].define(version: 2023_01_13_104231) do
     t.datetime "invitation_created_at", precision: nil
     t.datetime "invitation_sent_at", precision: nil
     t.datetime "invitation_accepted_at", precision: nil
+=======
+    t.datetime "reset_password_sent_at"
+    t.datetime "remember_created_at"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.string "confirmation_token"
+    t.datetime "confirmed_at"
+    t.datetime "confirmation_sent_at"
+    t.integer "permission"
+    t.string "invitation_token"
+    t.datetime "invitation_created_at"
+    t.datetime "invitation_sent_at"
+    t.datetime "invitation_accepted_at"
+>>>>>>> dev
     t.integer "invitation_limit"
     t.string "invited_by_type"
     t.bigint "invited_by_id"
     t.integer "invitations_count", default: 0
+<<<<<<< HEAD
     t.integer "account_id"
     t.index ["account_id"], name: "index_users_on_account_id"
   end
@@ -178,10 +295,39 @@ ActiveRecord::Schema[7.0].define(version: 2023_01_13_104231) do
   add_foreign_key "active_storage_attachments", "active_storage_blobs", column: "blob_id", name: "active_storage_attachments_blob_id_fkey"
   add_foreign_key "active_storage_variant_records", "active_storage_blobs", column: "blob_id", name: "active_storage_variant_records_blob_id_fkey"
   add_foreign_key "folders", "accounts"
+=======
+    t.boolean "active", default: true, null: false
+    t.date "deactivated_on"
+    t.boolean "email_enabled", default: true
+    t.integer "sign_in_count", default: 0, null: false
+    t.datetime "current_sign_in_at"
+    t.datetime "last_sign_in_at"
+    t.string "current_sign_in_ip"
+    t.string "last_sign_in_ip"
+    t.integer "role", default: 0
+    t.string "phone"
+    t.index ["confirmation_token"], name: "index_users_on_confirmation_token", unique: true
+    t.index ["email"], name: "index_users_on_email", unique: true
+    t.index ["invitation_token"], name: "index_users_on_invitation_token", unique: true
+    t.index ["invited_by_id"], name: "index_users_on_invited_by_id"
+    t.index ["invited_by_type", "invited_by_id"], name: "index_users_on_invited_by"
+    t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
+  end
+
+  add_foreign_key "active_storage_attachments", "active_storage_blobs", column: "blob_id"
+  add_foreign_key "active_storage_variant_records", "active_storage_blobs", column: "blob_id"
+>>>>>>> dev
   add_foreign_key "folders", "spaces"
   add_foreign_key "folders", "users"
   add_foreign_key "pinned_spaces", "spaces"
   add_foreign_key "pinned_spaces", "users"
+<<<<<<< HEAD
   add_foreign_key "spaces", "accounts"
   add_foreign_key "spaces", "users"
+=======
+  add_foreign_key "recent_surveys", "survey_surveys", column: "survey_surveys_id"
+  add_foreign_key "recent_surveys", "users"
+  add_foreign_key "spaces", "users"
+  add_foreign_key "survey_surveys", "folders"
+>>>>>>> dev
 end
