@@ -2,8 +2,8 @@ require "application_system_test_case"
 
 class SurveysTest < ApplicationSystemTestCase
   setup do
-    @user = users(:regular)
-    @survey = survey_surveys(:user)
+    @user = users(:member)
+    @survey = survey_surveys(:one)
     sign_in @user
   end
 
@@ -12,7 +12,7 @@ class SurveysTest < ApplicationSystemTestCase
   end
 
   def surveys_page_url
-    survey_questions_url(survey_id: @survey.id)
+    survey_url(@survey)
   end
 
   test "can show index if logged in" do
@@ -32,10 +32,19 @@ class SurveysTest < ApplicationSystemTestCase
     visit page_url
     find(id: dom_id(@survey)).click_link("Show")
     within "#survey-header" do
-      assert_text "Attempt"
-      assert_text "Clone"
+      assert_text @survey.name
+      assert_text "Actions"
     end
     take_screenshot
+  end
+
+  test "can attempt a survey" do
+    visit surveys_page_url
+    within "#survey-header" do
+      click_on "Actions"
+      click_on "Attempt"
+    end
+    assert_text "Enter Participant Details"
   end
 
   test "can create a new survey" do
@@ -67,6 +76,25 @@ class SurveysTest < ApplicationSystemTestCase
     assert_text "Survey Campaigning"
   end
 
+  test "can pin a survey" do
+    visit surveys_page_url
+    within "#survey-header" do
+      click_on "Actions"
+      click_on "Pin Survey"
+    end
+    assert_text "Survey has been pinned."
+  end
+
+  test "can unpin a survey" do
+    pinned = survey_surveys(:pinned)
+    visit survey_url(pinned)
+    within "#survey-header" do
+      click_on "Actions"
+      click_on "Unpin Survey"
+    end
+    assert_text "Survey has been unpinned."
+  end
+
   test "can not edit a survey with invalid name" do
     visit page_url
     find(id: dom_id(@survey)).click_link("Edit")
@@ -80,6 +108,7 @@ class SurveysTest < ApplicationSystemTestCase
     visit page_url
     find(id: dom_id(@survey)).click_link(@survey.name)
     within "#survey-header" do
+      click_on "Actions"
       page.accept_confirm do
         click_on "Clone"
       end
@@ -88,12 +117,37 @@ class SurveysTest < ApplicationSystemTestCase
     assert_text "#{@survey.name} (Copy)"
   end
 
-  test "can delete survey" do
-    visit page_url
-    page.accept_confirm do
-      find(id: dom_id(@survey)).click_link("Delete")
+  test "can restore archived survey" do
+    archived = survey_surveys(:archived)
+    visit archived_surveys_path
+    within "tr##{dom_id(archived)}" do
+      page.accept_confirm do
+        click_on "Unarchive"
+      end
     end
-    assert_no_text @survey.name
+    assert_text "Survey has been restored."
+  end
+
+  test "can archive a survey" do
+    visit page_url
+    survey = survey_surveys(:two)
+    within "tr##{dom_id(survey)}" do
+      page.accept_confirm do
+        click_on "Archive"
+      end
+    end
+    assert_text "Survey has been archived."
+  end
+
+  test "can delete archived survey" do
+    archived = survey_surveys(:archived)
+    visit archived_surveys_path
+    within "tr##{dom_id(archived)}" do
+      page.accept_confirm do
+        click_on "Delete"
+      end
+    end
+    assert_text "Survey has been deleted."
     take_screenshot
   end
 end
