@@ -2,9 +2,9 @@ require "application_system_test_case"
 
 class AttemptsTest < ApplicationSystemTestCase
   setup do
-    @user = users(:regular)
-    @survey = survey_surveys(:user)
-    @attempt = survey_attempts(:user)
+    @user = users(:member)
+    @survey = survey_surveys(:one)
+    @attempt = survey_attempts(:one)
     sign_in @user
   end
 
@@ -21,7 +21,6 @@ class AttemptsTest < ApplicationSystemTestCase
     take_screenshot
     assert_selector "h1", text: @survey.name
     assert_selector "div#survey-tabs", text: "Attempts"
-    assert_selector "div#assignees"
   end
 
   test "can not visit index if not logged in" do
@@ -30,42 +29,25 @@ class AttemptsTest < ApplicationSystemTestCase
     assert_selector "h1", text: "Sign in to your account"
   end
 
-  test "can attempt a survey" do
-    visit attempt_url
-    sleep(5)
+  test "can attempt a survey and view the attempt" do
+    visit page_url
+    within "#survey-header" do
+      click_on "Actions"
+      click_on "Attempt"
+    end
+    fill_in "name", with: "Test Attempt"
+    fill_in "email", with: "test_attempt@crownstack.com"
+    click_on "Start"
     click_on "Preview and Submit"
     take_screenshot
-    fill_in "survey_attempt_comment", with: "This is a comment"
+    fill_in "comment", with: "This is a comment"
     click_on "Submit"
     sleep(0.2)
     visit page_url
-    assert_selector "td", text: "Submitted"
-    find(id: dom_id(@attempt)).click_link("View")
-    take_screenshot
-    assert_selector "h4", text: "This is a comment"
-  end
-
-  test "can view  attempt" do
-    visit page_url
-    take_screenshot
-    @attempt = @survey.attempts.first
-    if find("tr", id: dom_id(@attempt))
-      find("tr", id: dom_id(@attempt)).click_link("View")
-      assert_selector "h1", text: @survey.name
-      assert_selector "p", text: "Participant"
-      assert_selector "h3", text: "Score"
+    attempt = @survey.attempts.last
+    within "tr", id: dom_id(attempt) do
+      click_on @user.decorate.display_name
     end
-    take_screenshot
-  end
-
-  test "can delete attempt" do
-    visit page_url
-    take_screenshot
-    @attempt = @survey.attempts.first
-    if find("tr", id: dom_id(@attempt))
-      find("tr", id: dom_id(@attempt)).click_link("Delete")
-      assert_no_selector dom_id(@attempt), text: @attempt.survey
-    end
-    take_screenshot
+    assert_selector "h1", text: attempt.participant.name
   end
 end
