@@ -2,13 +2,13 @@ require "application_system_test_case"
 
 class QuestionsTest < ApplicationSystemTestCase
   setup do
-    @user = users(:regular)
-    @survey = survey_surveys(:user)
+    @user = users(:member)
+    @survey = survey_surveys(:one)
     sign_in @user
   end
 
   def page_url
-    survey_questions_url(survey_id: @survey.id)
+    survey_path(@survey)
   end
 
   test "can show index if logged in" do
@@ -29,7 +29,6 @@ class QuestionsTest < ApplicationSystemTestCase
     visit page_url
     fill_in "survey_question_text", with: "This is a question"
     fill_in "survey_question_description", with: "This is a sample Question Description"
-    select survey_question_categories(:one).name, from: "survey_question_question_category_id"
     click_on "Add Question"
     assert_selector "p", text: "Question was added successfully."
     take_screenshot
@@ -45,9 +44,9 @@ class QuestionsTest < ApplicationSystemTestCase
   test "can edit a question" do
     visit page_url
     @question = @survey.questions.first
-    find("li", id: "#{@question.id}").click_link("Edit")
-    fill_in "survey_question_text", with: "question 1"
-    click_on "Save"
+    find("li", id: dom_id(@question)).click_link("Edit")
+    fill_in "survey_question_text", with: "question 1 new"
+    click_on "Edit Question"
     assert_text "Question 1"
     assert_no_text "Save"
     take_screenshot
@@ -56,16 +55,18 @@ class QuestionsTest < ApplicationSystemTestCase
   test "can not edit a survey with invalid name" do
     visit page_url
     @question = @survey.questions.first
-    find("li", id: "#{@question.id}").click_link("Edit")
+    find("li", id: dom_id(@question)).click_link("Edit")
     fill_in "survey_question_text", with: ""
-    click_on "Save"
+    click_on "Edit Question"
     take_screenshot
   end
 
   test "can delete a question" do
     visit page_url
     @question = @survey.questions.first
-    find("li", id: "#{@question.id}").click_link("Delete")
+    page.accept_confirm do
+      find("li", id: dom_id(@question)).click_link("Delete")
+    end
     take_screenshot
     assert_no_text @question.decorate.display_text
   end
